pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "src"
version = "0.0.1"
description = "My first slm, in honor to francesco damiano di gregorio one of the craziest person that I know."
authors = [
  { name = "andrea porcelli" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "datasets>=4.8.4",
    "hydra-core>=1.3.2",
    "pip",
    "python-dotenv",
    "ruff",
    "torch>=2.11.0",
    "torchtune>=0.6.1",
    "transformers>=5.4.0",
    "wandb>=0.25.1",
]
requires-python = "~=3.13.0"


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

